+++ src/core/client/views/atm/atm.ts
@@ -1,5 +1,6 @@
 import * as alt from 'alt-client';
 import * as native from 'natives';
+
 import { SYSTEM_EVENTS } from '../../../shared/enums/system';
 import { Character } from '../../../shared/interfaces/Character';
 import { LOCALE_KEYS } from '../../../shared/locale/languages/keys';