+++ src/core/plugins/interactionExample/index.ts
@@ -5,18 +5,24 @@ import { InteractionController } from '../../server/systems/interaction';
 alt.on('interaction:DoSomething', doSomeInteracting);
 
 function GenerateInteractions() {
-    InteractionController.addInteraction(
-        'interaction:DoSomething',
-        { x: 402.397308, y: -1029.67, z: 29.34688 },
-        3,
-        'Do the thing!',
-        null,
-        true
-    );
+    InteractionController.add({
+        event: { eventName: `interaction:DoSomething`, isServer: true },
+        type: 'interaction:DoSomething',
+        position: { x: 402.397308, y: -1029.67, z: 29.34688 },
+        shortDesc: 'Browse Heist',
+        blip: {
+            sprite: 112,
+            color: 7,
+            shortRange: true,
+            text: 'Do Something!',
+            pos: { x: 402.397308, y: -1029.67, z: 29.34688 },
+            scale: 1
+        }
+    });
 }
 
 function doSomeInteracting(player: alt.Player) {
     playerFuncs.emit.message(player, 'Nice!');
 }
 
-GenerateInteractions();
+// GenerateInteractions();