+++ src/core/shared/interfaces/Help.ts
@@ -3,6 +3,6 @@ import { Vector3 } from './Vector';
 export interface Help {
     position: Vector3;
     key?: number | null;
-    shortDesc?: string | null;
-    longDesc?: string | null;
+    desc?: string | null;
+    expiration?: number;
 }