+++ src/core/shared/interfaces/Interaction.ts
@@ -1,10 +1,12 @@
 import { Blip } from './Blip';
+import { EventCall } from './EventCall';
+import { Vector3 } from './Vector';
 
 export interface Interaction {
+    description?: string;
+    position?: Vector3;
     identifier?: string;
-    text: string;
-    blip: Blip;
-    eventName?: string;
-    maxRadius?: number;
-    isServer?: boolean;
+    type?: string;
+    blip?: Blip;
+    event?: EventCall;
 }